Add files via upload
**Algorithm for Calculating Students' Average and Status**

This algorithm is designed to calculate the average grades of four students and determine whether each student is approved or failed based on the entered grades. The program prompts the user to input the name and two grades for each student, calculates the average, and then displays the student's name along with their average. If the average is less than or equal to 4, the student is considered failed; otherwise, the student is approved. The algorithm uses a loop to process data for all students and displays the results in an organized manner.